Gotten started on the spreadsheet test
