+++ static/redesign/tests/spec-runner.js
@@ -23,7 +23,8 @@ require(['boot'], function () {
         'spec/views/data-manager-test.js',
         'spec/views/legend-layer-entry-test.js',
         'spec/views/audio-player-test.js',
-        'spec/views/create-form-test.js'
+        'spec/views/create-form-test.js',
+        'spec/views/spreadsheet-test.js'
     ];
 
     require(specs, function () {
